Start Training finish & improve code

+++ app/src/main/java/com/example/desafiolealapps/ui/activities/EditTrainingActivity.kt
@@ -5,10 +5,13 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import android.view.View
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.desafiolealapps.data.ItemExercise
+import com.example.desafiolealapps.data.ItemTraining
 import com.example.desafiolealapps.databinding.ActivityEditTrainingBinding
 import com.example.desafiolealapps.ui.adapters.AdapterExerciseList
+import com.example.desafiolealapps.ui.adapters.AdapterStartExerciseList
 import com.google.firebase.Firebase
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.auth
@@ -17,10 +20,12 @@ import com.google.firebase.firestore.FirebaseFirestore
 class EditTrainingActivity : AppCompatActivity() {
     lateinit var binding: ActivityEditTrainingBinding
     lateinit var trainingId: String
-    lateinit var currentTrainingId: String
     private lateinit var auth: FirebaseAuth
     private lateinit var adapter: AdapterExerciseList
     private val db = FirebaseFirestore.getInstance()
+    private lateinit var trainingName: String
+    private lateinit var trainingDescription: String
+    private lateinit var trainingDays: String
 
     companion object {
         private const val EXERCISE_REQUEST_CODE = 321
@@ -33,14 +38,14 @@ class EditTrainingActivity : AppCompatActivity() {
 
         auth = Firebase.auth
         trainingId = intent.getStringExtra("trainingId").toString()
-        setupContentView()
-        val trainingName = intent.getStringExtra("trainingName")
-        val trainingDescription = intent.getStringExtra("trainingDescription")
-        val trainingTime = intent.getStringExtra("trainingTime")
-
+        trainingName = intent.getStringExtra("trainingName").toString()
+        trainingDescription = intent.getStringExtra("trainingDescription").toString()
+        trainingDays = intent.getStringExtra("trainingDays").toString()
         binding.editTextTrainingName.setText(trainingName)
-        binding.editTextTrainingTime.setText(trainingTime)
+        setCheckedDays(trainingDays)
         binding.editTextTrainingDescription.setText(trainingDescription)
+        setupContentView()
+
     }
 
     private fun setupContentView() {
@@ -49,22 +54,79 @@ class EditTrainingActivity : AppCompatActivity() {
         setupBackButton()
         loadExercisesFromDatabase()
         setupGoToCreateExerciseButton()
+        setupGoToStartTrainingButton()
         setupRecyclerViewExercises()
 
         if (intent.getStringExtra("trainingId") != null) {
             enableCreateExerciseButton()
-            setupUpdateTrainingButton()
             setupDeleteButton()
         } else {
             binding.delete.visibility = View.GONE
-            setupCreateTrainingButton()
         }
     }
 
-    private fun setupBackButton(){
+    private fun setupBackButton() {
         binding.back.setOnClickListener {
+
+            val currentTrainingDays = getSelectedDaysString()
+            val currentTrainingDescription = binding.editTextTrainingDescription.text.toString()
+            val currentTrainingName = binding.editTextTrainingName.text.toString()
+
+            if (trainingDays != currentTrainingDays ||
+                trainingDescription != currentTrainingDescription ||
+                trainingName != currentTrainingName
+            ) {
+                showConfirmationDialog()
+            } else {
+                onBackPressed()
+            }
+        }
+    }
+
+    private fun showConfirmationDialog() {
+        val builder = AlertDialog.Builder(this)
+        builder.setTitle("Confirmação")
+        builder.setMessage("Deseja salvar as alterações?")
+
+        builder.setPositiveButton("Sim") { dialog, which ->
+            updateTrainingButton()
+            setResult(RESULT_OK)
+        }
+
+        builder.setNegativeButton("Não") { dialog, which ->
             onBackPressed()
         }
+
+        val dialog = builder.create()
+        dialog.show()
+    }
+
+    fun getSelectedDaysString(): String {
+        val selectedDays = mutableListOf<String>()
+
+        if (binding.checkboxSunday.isChecked) {
+            selectedDays.add("dom")
+        }
+        if (binding.checkboxMonday.isChecked) {
+            selectedDays.add("seg")
+        }
+        if (binding.checkboxTuesday.isChecked) {
+            selectedDays.add("ter")
+        }
+        if (binding.checkboxWednesday.isChecked) {
+            selectedDays.add("qua")
+        }
+        if (binding.checkboxThursday.isChecked) {
+            selectedDays.add("qui")
+        }
+        if (binding.checkboxFriday.isChecked) {
+            selectedDays.add("sex")
+        }
+        if (binding.checkboxSaturday.isChecked) {
+            selectedDays.add("sáb")
+        }
+
+        return selectedDays.joinToString(", ")
     }
 
     private fun setupRecyclerViewExercises() {
@@ -78,7 +140,40 @@ class EditTrainingActivity : AppCompatActivity() {
             }
         })
 
+        adapter.setOnItemLongClickListener(object : AdapterExerciseList.OnItemLongClickListener {
+            override fun onItemLongClicked(exercise: ItemExercise): Boolean {
+                val builder = AlertDialog.Builder(this@EditTrainingActivity)
+                builder.setTitle("Observação")
+                builder.setMessage(exercise.exerciseObservation)
+                builder.setPositiveButton("OK") { _, _ ->
+                }
+
+                val dialog = builder.create()
+                dialog.show()
+                return true
+            }
+        })
+
         loadExercisesFromDatabase()
+    }
+
+    private fun setCheckedDays(selectedDays: String?) {
+
+        val daysArray = selectedDays?.split(", ")?.toTypedArray()
+
+        if (daysArray != null) {
+            for (day in daysArray) {
+                when (day) {
+                    "dom" -> binding.checkboxSunday.isChecked = true
+                    "seg" -> binding.checkboxMonday.isChecked = true
+                    "ter" -> binding.checkboxTuesday.isChecked = true
+                    "qua" -> binding.checkboxWednesday.isChecked = true
+                    "qui" -> binding.checkboxThursday.isChecked = true
+                    "sex" -> binding.checkboxFriday.isChecked = true
+                    "sáb" -> binding.checkboxSaturday.isChecked = true
+                }
+            }
+        }
     }
 
     private fun loadExercisesFromDatabase() {
@@ -96,12 +191,14 @@ class EditTrainingActivity : AppCompatActivity() {
                         val exerciseName = document.getString("exerciseName") ?: ""
                         val exerciseTime = document.getString("exerciseTime") ?: ""
                         val exerciseObservation = document.getString("exerciseObservation") ?: ""
+                        val exerciseRepetition = document.getString("exerciseRepetition") ?: ""
 
                         val newItem = ItemExercise(
                             document.id,
                             exerciseName,
                             exerciseObservation,
-                            exerciseTime
+                            exerciseTime,
+                            exerciseRepetition
                         )
 
                         exercisesList.add(newItem)
@@ -125,17 +222,35 @@ class EditTrainingActivity : AppCompatActivity() {
         }
     }
 
+    private fun setupGoToStartTrainingButton() {
+        val currentTrainingId = trainingId
+        val intent = Intent(this, StartTrainingActivity::class.java)
+        intent.putExtra("trainingId", currentTrainingId)
+        binding.startTraining.setOnClickListener {
+            startActivityForResult(intent, EXERCISE_REQUEST_CODE)
+        }
+    }
+
     private fun setupDeleteButton() {
         binding.delete.visibility = View.VISIBLE
         binding.delete.setOnClickListener {
             val trainingId = intent.getStringExtra("trainingId") ?: return@setOnClickListener
-            val currentUser = auth.currentUser
 
-            currentUser?.uid?.let { userId ->
-                deleteTraining(userId, trainingId)
-            } ?: run {
-                Log.e("EditTrainingActivity", "User is null")
+            val builder = AlertDialog.Builder(this)
+            builder.setTitle("Confirmação")
+            builder.setMessage("Deseja realmente deletar este treino?")
+            builder.setPositiveButton("Sim") { dialog, which ->
+                val currentUser = auth.currentUser
+
+                currentUser?.uid?.let { userId ->
+                    deleteTraining(userId, trainingId)
+                } ?: run {
+                    Log.e("EditTrainingActivity", "User is null")
+                }
             }
+            builder.setNegativeButton("Não") { dialog, which -> }
+            val dialog = builder.create()
+            dialog.show()
         }
     }
 
@@ -158,26 +273,24 @@ class EditTrainingActivity : AppCompatActivity() {
             }
     }
 
-    private fun setupUpdateTrainingButton() {
-        binding.createTrainingButton.setOnClickListener {
-            val trainingId = intent.getStringExtra("trainingId") ?: return@setOnClickListener
+    private fun updateTrainingButton() {
+            val trainingId = intent.getStringExtra("trainingId") ?: return
             val trainingName = binding.editTextTrainingName.text.toString()
             val trainingDescription = binding.editTextTrainingDescription.text.toString()
-            val trainingTime = binding.editTextTrainingTime.text.toString()
+            val trainingDays = getSelectedDaysString()
 
             val currentUser = auth.currentUser
             currentUser?.uid?.let { userId ->
                 val trainingMap = hashMapOf(
                     "trainingName" to trainingName,
                     "trainingDescription" to trainingDescription,
-                    "trainingTime" to trainingTime
+                    "trainingDays" to trainingDays
                 )
 
                 updateTraining(userId, trainingId, trainingMap)
             } ?: run {
                 Log.e("EditTrainingActivity", "User is null")
             }
-        }
     }
 
     private fun updateTraining(userId: String, trainingId: String, trainingMap: Map<String, Any>) {
@@ -189,59 +302,13 @@ class EditTrainingActivity : AppCompatActivity() {
                 if (task.isSuccessful) {
                     Log.d("db", "Success updating training! Document ID: $trainingId")
                     setResult(RESULT_OK)
-                    enableCreateExerciseButton()
+                    onBackPressed()
                 } else {
                     Log.e("db", "Error updating training", task.exception)
                 }
             }
             .addOnFailureListener { exception ->
                 Log.e("db", "Error updating training", exception)
-            }
-    }
-
-    private fun setupCreateTrainingButton() {
-        binding.createTrainingButton.setOnClickListener {
-            val trainingName = binding.editTextTrainingName.text.toString()
-            val trainingDescription = binding.editTextTrainingDescription.text.toString()
-            val trainingTime = binding.editTextTrainingTime.text.toString()
-
-            val currentUser = auth.currentUser
-            currentUser?.uid?.let { userId ->
-                val trainingMap = hashMapOf(
-                    "trainingName" to trainingName,
-                    "trainingDescription" to trainingDescription,
-                    "trainingTime" to trainingTime
-                )
-
-                createTraining(userId, trainingMap)
-            } ?: run {
-                Log.e("EditTrainingActivity", "User is null")
-            }
-        }
-    }
-
-    private fun createTraining(userId: String, trainingMap: Map<String, Any>) {
-        db.collection("Users").document(userId)
-            .collection("trainings")
-            .add(trainingMap)
-            .addOnCompleteListener { task ->
-                if (task.isSuccessful) {
-                    val newTrainingId = task.result?.id
-                    Log.d("db", "Success creating training! Document ID: $newTrainingId")
-
-                    if (newTrainingId != null) {
-                        setResult(RESULT_OK)
-                        trainingId = newTrainingId
-                        enableCreateExerciseButton()
-                    } else {
-                        Log.e("db", "Error: Document ID is null")
-                    }
-                } else {
-                    Log.e("db", "Error creating training", task.exception)
-                }
-            }
-            .addOnFailureListener { exception ->
-                Log.e("db", "Error creating training", exception)
             }
     }
 
@@ -259,6 +326,7 @@ class EditTrainingActivity : AppCompatActivity() {
         intent.putExtra("exerciseName", exercise.exerciseName)
         intent.putExtra("exerciseTime", exercise.exerciseTime)
         intent.putExtra("exerciseObservation", exercise.exerciseObservation)
+        intent.putExtra("exerciseRepetition", exercise.exerciseRepetition)
         startActivityForResult(intent, EXERCISE_REQUEST_CODE)
     }
